Added an outside scene and main menu.
I added an outside scene where I will put the shop and the shop keeper.

I added a circular transition when switching between scenes.

I added a Main Menu with two options. 'Play' which will bring you to the beginning of the game, and 'Quit' which will close the game.