Final Build
This is the final Push I have for this game.

I added a shopkeeper that you can talk to and open up the shop menu. In the shop menu you can choose to purchase shirts or pants. Once purchased from the shopkeeper you will be able to equip the clothing in the same menu.